feat(billing): host quota tracker runtime with validated inputs
